resharper again

+++ DamageMeter.AutoUpdate/UpdateManager.cs
@@ -29,10 +29,7 @@ namespace DamageMeter.AutoUpdate
                 while (directory != null)
                 {
                     var resourceDirectory = Path.Combine(directory, @"resources\");
-                    if (Directory.Exists(resourceDirectory))
-                    {
-                        return resourceDirectory;
-                    }
+                    if (Directory.Exists(resourceDirectory)) { return resourceDirectory; }
                     directory = Path.GetDirectoryName(directory);
                 }
                 throw new InvalidOperationException("Could not find the resource directory");
@@ -48,16 +45,8 @@ namespace DamageMeter.AutoUpdate
         public static void ReadDbVersion()
         {
             var version = Path.Combine(ResourcesDirectory, "head");
-            try
-            {
-                if (File.Exists(version))
-                {
-                    Version = Version + "." + File.ReadLines(version).FirstOrDefault()?.Remove(7);
-                }
-            }
-            catch
-            {
-            } // ignore bad head
+            try { if (File.Exists(version)) { Version = Version + "." + File.ReadLines(version).FirstOrDefault()?.Remove(7); } }
+            catch { } // ignore bad head
         }
 
         public static bool Update()
@@ -70,12 +59,8 @@ namespace DamageMeter.AutoUpdate
         {
             using (var client = new WebClient())
             {
-                var compressed =
-                    client.OpenRead(new Uri("https://neowutran.ovh/updates/ShinraMeterV/" + file.Key + ".zip"));
-                if (compressed == null)
-                {
-                    return true;
-                }
+                var compressed = client.OpenRead(new Uri("https://neowutran.ovh/updates/ShinraMeterV/" + file.Key + ".zip"));
+                if (compressed == null) { return true; }
                 new ZipArchive(compressed).Entries[0].ExtractToFile(ExecutableDirectory + @"\tmp\release\" + file.Key);
             }
             return FileHash(ExecutableDirectory + @"\tmp\release\" + file.Key) != file.Value;
@@ -83,11 +68,8 @@ namespace DamageMeter.AutoUpdate
 
         internal static Dictionary<string, string> ReadHashFile(string file, string addPath = "")
         {
-            return File.ReadAllLines(file)
-                .Where(s => !string.IsNullOrWhiteSpace(s))
-                .Select(s => s.Split(new[] {" *"}, StringSplitOptions.None))
-                .Select(parts => new KeyValuePair<string, string>(addPath + parts[1], parts[0]))
-                .ToDictionary(x => x.Key, x => x.Value);
+            return File.ReadAllLines(file).Where(s => !string.IsNullOrWhiteSpace(s)).Select(s => s.Split(new[] {" *"}, StringSplitOptions.None))
+                .Select(parts => new KeyValuePair<string, string>(addPath + parts[1], parts[0])).ToDictionary(x => x.Key, x => x.Value);
         }
 
         private static bool HashedUpdate()
@@ -95,14 +77,9 @@ namespace DamageMeter.AutoUpdate
             DestroyDownloadDirectory();
             Directory.CreateDirectory(ExecutableDirectory + @"\tmp\release\");
             var fileList = _latest.Except(_hashes).ToList();
-            if (!fileList.Any())
-            {
-                return false;
-            }
-            File.WriteAllLines(ExecutableDirectory + @"\tmp\ShinraMeterV.sha1",
-                _latest.Select(x => x.Value + " *" + x.Key));
-            fileList.Where(x => x.Key.Contains('\\'))
-                .Select(x => Path.GetDirectoryName(ExecutableDirectory + @"\tmp\release\" + x.Key)).Distinct().ToList()
+            if (!fileList.Any()) { return false; }
+            File.WriteAllLines(ExecutableDirectory + @"\tmp\ShinraMeterV.sha1", _latest.Select(x => x.Value + " *" + x.Key));
+            fileList.Where(x => x.Key.Contains('\\')).Select(x => Path.GetDirectoryName(ExecutableDirectory + @"\tmp\release\" + x.Key)).Distinct().ToList()
                 .ForEach(x => Directory.CreateDirectory(x));
             var badhash = false;
             fileList.ForEach(x => badhash = badhash || GetDiff(x));
@@ -113,13 +90,9 @@ namespace DamageMeter.AutoUpdate
             }
             if (File.Exists(ExecutableDirectory + @"\tmp\release\Autoupdate.exe"))
             {
-                File.Copy(ExecutableDirectory + @"\tmp\release\Autoupdate.exe",
-                    ExecutableDirectory + @"\tmp\Autoupdate.exe");
+                File.Copy(ExecutableDirectory + @"\tmp\release\Autoupdate.exe", ExecutableDirectory + @"\tmp\Autoupdate.exe");
             }
-            else
-            {
-                File.Copy(ExecutableDirectory + @"\Autoupdate.exe", ExecutableDirectory + @"\tmp\Autoupdate.exe");
-            }
+            else { File.Copy(ExecutableDirectory + @"\Autoupdate.exe", ExecutableDirectory + @"\tmp\Autoupdate.exe"); }
             Process.Start(ExecutableDirectory + @"\tmp\Autoupdate.exe", "pass");
             return true;
         }
@@ -136,10 +109,7 @@ namespace DamageMeter.AutoUpdate
 
         private static void DestroyDownloadDirectory()
         {
-            if (!Directory.Exists(ExecutableDirectory + @"\tmp\"))
-            {
-                return;
-            }
+            if (!Directory.Exists(ExecutableDirectory + @"\tmp\")) { return; }
             Directory.Delete(ExecutableDirectory + @"\tmp\", true);
         }
 
@@ -148,13 +118,7 @@ namespace DamageMeter.AutoUpdate
             Directory.CreateDirectory(targetDir);
             foreach (var file in Directory.GetFiles(sourceDir))
             {
-                if (Path.GetFileName(file) == "ShinraLauncher.exe")
-                {
-                    if (File.Exists(Path.Combine(targetDir, Path.GetFileName(file))))
-                    {
-                        continue;
-                    }
-                }
+                if (Path.GetFileName(file) == "ShinraLauncher.exe") { if (File.Exists(Path.Combine(targetDir, Path.GetFileName(file)))) { continue; } }
                 File.Copy(file, Path.Combine(targetDir, Path.GetFileName(file)), true);
             }
 
@@ -171,29 +135,16 @@ namespace DamageMeter.AutoUpdate
 
         internal static void DestroyRelease()
         {
-            Array.ForEach(
-                Directory.GetFiles(ExecutableDirectory + @"\..\..\").Where(t => !t.EndsWith("ShinraLauncher.exe"))
-                    .ToArray(), File.Delete);
+            Array.ForEach(Directory.GetFiles(ExecutableDirectory + @"\..\..\").Where(t => !t.EndsWith("ShinraLauncher.exe")).ToArray(), File.Delete);
             Array.ForEach(Directory.GetFiles(ExecutableDirectory + @"\..\..\resources\"), File.Delete);
-            foreach (var s in Directory.GetDirectories(ExecutableDirectory + @"\..\..\")
-                .Where(t => !(t.EndsWith("resources") || t.EndsWith("tmp"))))
+            foreach (var s in Directory.GetDirectories(ExecutableDirectory + @"\..\..\").Where(t => !(t.EndsWith("resources") || t.EndsWith("tmp"))))
             {
-                if (Directory.Exists(s))
-                {
-                    Directory.Delete(s, true);
-                }
+                if (Directory.Exists(s)) { Directory.Delete(s, true); }
             }
-            if (!Directory.Exists(ExecutableDirectory + @"\..\..\resources\"))
+            if (!Directory.Exists(ExecutableDirectory + @"\..\..\resources\")) { return; }
+            foreach (var s in Directory.GetDirectories(ExecutableDirectory + @"\..\..\resources\").Where(t => !t.EndsWith("config") && !t.EndsWith("sound")))
             {
-                return;
-            }
-            foreach (var s in Directory.GetDirectories(ExecutableDirectory + @"\..\..\resources\")
-                .Where(t => !t.EndsWith("config") && !t.EndsWith("sound")))
-            {
-                if (Directory.Exists(s))
-                {
-                    Directory.Delete(s, true);
-                }
+                if (Directory.Exists(s)) { Directory.Delete(s, true); }
             }
             Console.WriteLine("Resources directory destroyed");
         }
@@ -203,19 +154,16 @@ namespace DamageMeter.AutoUpdate
             Parallel.ForEach(Directory.GetDirectories(parentDirectory), directory =>
             {
                 DeleteEmptySubdirectories(directory);
-                if (!Directory.EnumerateFileSystemEntries(directory).Any())
-                {
-                    Directory.Delete(directory, false);
-                }
+                if (!Directory.EnumerateFileSystemEntries(directory).Any()) { Directory.Delete(directory, false); }
             });
         }
 
         internal static void CleanupRelease(Dictionary<string, string> hashes)
         {
-            Array.ForEach(Directory.GetFiles(ExecutableDirectory + @"\..\", "*", SearchOption.AllDirectories)
-                    .Where(t => !(t.Contains(@"\config\") || t.Contains(@"\..\tmp\") || t.Contains(@"\sound\") ||
-                                  t.EndsWith("ShinraLauncher.exe") || hashes.ContainsKey(t))).ToArray()
-                , x =>
+            Array.ForEach(
+                Directory.GetFiles(ExecutableDirectory + @"\..\", "*", SearchOption.AllDirectories)
+                    .Where(t => !(t.Contains(@"\config\") || t.Contains(@"\..\tmp\") || t.Contains(@"\sound\") || t.EndsWith("ShinraLauncher.exe") ||
+                                  hashes.ContainsKey(t))).ToArray(), x =>
                 {
                     File.Delete(x);
                     Console.WriteLine(x);
@@ -228,22 +176,14 @@ namespace DamageMeter.AutoUpdate
         {
             using (var client = new WebClient())
             {
-                var compressed = await client
-                    .OpenReadTaskAsync(new Uri("http://diclah.com/~yukikoo/ShinraMeterV.sha1.zip"))
-                    .ConfigureAwait(false);
-                if (compressed == null)
-                {
-                    return true;
-                }
+                var compressed = await client.OpenReadTaskAsync(new Uri("http://diclah.com/~yukikoo/ShinraMeterV.sha1.zip")).ConfigureAwait(false);
+                if (compressed == null) { return true; }
                 using (var stream = new MemoryStream())
                 {
                     new ZipArchive(compressed).Entries[0].Open().CopyTo(stream);
-                    _latest = Encoding.UTF8.GetString(stream.ToArray())
-                        .Split(new[] {"\r\n", "\r", "\n"}, StringSplitOptions.None)
-                        .Where(s => !string.IsNullOrWhiteSpace(s))
-                        .Select(s => s.Split(new[] {" *"}, StringSplitOptions.None))
-                        .Select(parts => new KeyValuePair<string, string>(parts[1], parts[0]))
-                        .ToDictionary(x => x.Key, x => x.Value);
+                    _latest = Encoding.UTF8.GetString(stream.ToArray()).Split(new[] {"\r\n", "\r", "\n"}, StringSplitOptions.None)
+                        .Where(s => !string.IsNullOrWhiteSpace(s)).Select(s => s.Split(new[] {" *"}, StringSplitOptions.None))
+                        .Select(parts => new KeyValuePair<string, string>(parts[1], parts[0])).ToDictionary(x => x.Key, x => x.Value);
                 }
             }
             CurrentHash();
@@ -266,10 +206,10 @@ namespace DamageMeter.AutoUpdate
         private static void CurrentHash()
         {
             _hashes = new Dictionary<string, string>();
-            Array.ForEach(Directory.GetFiles(ExecutableDirectory, "*", SearchOption.AllDirectories).Where(t =>
-                !t.EndsWith("ShinraLauncher.exe") && !t.Contains(@"\tmp\") && !t.Contains(@"\config\") &&
-                !t.Contains(@"\sound\") && !t.EndsWith("error.log")
-            ).ToArray(), x => _hashes.Add(x.Replace(ExecutableDirectory + "\\", ""), FileHash(x)));
+            Array.ForEach(
+                Directory.GetFiles(ExecutableDirectory, "*", SearchOption.AllDirectories)
+                    .Where(t => !t.EndsWith("ShinraLauncher.exe") && !t.Contains(@"\tmp\") && !t.Contains(@"\config\") && !t.Contains(@"\sound\") &&
+                                !t.EndsWith("error.log")).ToArray(), x => _hashes.Add(x.Replace(ExecutableDirectory + "\\", ""), FileHash(x)));
         }
     }
 }
